Add UltiSnips
UltiSnips is a snippet manager, but it also requires a directory to
store its so-called "snippets". This installs UltiSnips into the vim
configuration and also creates a path within .vim to store those
snippets. Perhaps in the future it will make sense to make a separate
repo (or a flake) out of those.